Use a real fix for the incorrect edge issue
The new pixel in the infinite border depends on what the current infinite border is. If the current infinite border is dark, during enhancement the window it will produce is:
...
...
...
Which corresponds to the first index of the algorithm. If it is light the window produced is:
###
###
###
Which corresponds to the last index of the algorithm.

When creating the initial image, set the infinite border pixel to dark.
On every enhancement step, use the current infinite border pixel to populate window coordinates that are out of range. Then when returning the enhanced image check what the new infinite border pixel produced by the algorithm will be